Refactor DbSite into its own file
