+++ src/test_s4solution.cpp
@@ -12,6 +12,8 @@
 #include "gtest/gtest.h"
 
 #include "processinstance.h"
+#include "db_site.h"
+
 #include "memcachedstore.h"
 #include "astaire_aor_store.h"
 #include "s4.h"
@@ -22,202 +24,6 @@
 #include <stdio.h>
 #include <thread>
 #include <boost/filesystem.hpp>
-
-static const int MEMCACHED_PORT = 33333;
-static const int ROGERS_PORT = 11311;
-static const int CHRONOS_PORT = 7253;
-
-class DbSite
-{
-public:
-  DbSite(int index, const std::string& dir);
-  virtual ~DbSite();
-
-  void create_and_start_memcached_instances(int memcached_instances);
-  void create_and_start_rogers_instances(int rogers_instances);
-  void create_and_start_chronos_instances(int chronos_instances);
-
-  std::vector<std::string> get_chronos_ips();
-  std::vector<std::string> get_rogers_ips();;
-
-  bool wait_for_instances();
-
-private:
-  /// Utility function to get at one of the site IPs.
-  std::string site_ip(int index);
-
-  /// Index of this site. Each site must have a different index.
-  int _site_index;
-
-  /// Directory this site may use for storing config files, log files, etc.
-  std::string _site_dir;
-
-  /// Use shared pointers for managing the instances so that the memory gets
-  /// freed when the vector is cleared.
-  std::vector<std::shared_ptr<MemcachedInstance>> _memcached_instances;
-  std::vector<std::shared_ptr<RogersInstance>> _rogers_instances;
-  std::vector<std::shared_ptr<ChronosInstance>> _chronos_instances;
-};
-
-
-DbSite::DbSite(int index, const std::string& dir) :
-  _site_index(index), _site_dir(dir)
-{
-  boost::filesystem::create_directory(_site_dir);
-}
-
-DbSite::~DbSite()
-{
-  _memcached_instances.clear();
-  _rogers_instances.clear();
-  _chronos_instances.clear();
-
-  if (remove("cluster_settings") != 0)
-  {
-    perror("remove cluster_settings");
-  }
-
-  boost::filesystem::remove_all(_site_dir);
-}
-
-
-std::string DbSite::site_ip(int index)
-{
-  return "127.0." + std::to_string(_site_index) + "." + std::to_string(index);
-}
-
-/// Creates and starts up the specified number of memcached instances. Also
-/// sets up the appropriate cluster_settings file (which just contains a
-/// single line of the form:
-///
-///   servers=127.0.0.1:33333,127.0.0.2:33333,...
-///
-void DbSite::create_and_start_memcached_instances(int memcached_instances)
-{
-  std::ofstream cluster_settings("cluster_settings");
-
-  for (int ii = 0; ii < memcached_instances; ++ii)
-  {
-    if (ii == 0)
-    {
-      cluster_settings << "servers=";
-    }
-    else
-    {
-      cluster_settings << ",";
-    }
-
-    // Each instance should listen on a new IP address.
-    std::string ip = site_ip(ii + 1);
-    _memcached_instances.emplace_back(new MemcachedInstance(ip, MEMCACHED_PORT));
-    _memcached_instances.back()->start_instance();
-
-    cluster_settings << ip << ":" << std::to_string(MEMCACHED_PORT);
-  }
-
-  cluster_settings.close();
-}
-
-/// Creates and starts up the specified number of Rogers instances.
-void DbSite::create_and_start_rogers_instances(int rogers_instances)
-{
-  for (int ii = 0; ii < rogers_instances; ++ii)
-  {
-    _rogers_instances.emplace_back(new RogersInstance(site_ip(ii + 1), ROGERS_PORT));
-    _rogers_instances.back()->start_instance();
-  }
-}
-
-/// Creates and starts up the specified number of chronos instances.
-void DbSite::create_and_start_chronos_instances(int chronos_instances)
-{
-  // Create directory to hold chronos config and logs.
-  std::string chronos_dir = _site_dir + "/chronos";
-  boost::filesystem::create_directory(chronos_dir);
-
-  // Create the chronos config that is common across all nodes in the cluster.
-  std::string cluster_conf_file = chronos_dir + "/chronos_cluster.conf";
-  std::ofstream conf;
-  conf.open(cluster_conf_file);
-  conf << "[cluster]\n";
-
-  for (int ii = 0; ii < chronos_instances; ++ii)
-  {
-    std::string ip = site_ip(ii + 1);
-    conf << "node = " << ip << ":" << std::to_string(CHRONOS_PORT) << "\n";
-
-    std::string dir = chronos_dir + "/instance" + std::to_string(ii + 1);
-    _chronos_instances.emplace_back(new ChronosInstance(ip,
-                                                        CHRONOS_PORT,
-                                                        dir,
-                                                        cluster_conf_file));
-    _chronos_instances.back()->start_instance();
-  }
-
-  conf.close();
-}
-
-/// Wait for all existing memcached, rogers and chronos instances to come up by
-/// checking they're listening on the correct ports. Returns false if any of the
-/// instances fail to come up.
-bool DbSite::wait_for_instances()
-{
-  bool success = true;
-
-  for (const std::shared_ptr<MemcachedInstance>& inst : _memcached_instances)
-  {
-    success = inst->wait_for_instance();
-
-    if (!success)
-    {
-      return success;
-    }
-  }
-
-  for (const std::shared_ptr<RogersInstance>& inst : _rogers_instances)
-  {
-    success = inst->wait_for_instance();
-
-    if (!success)
-    {
-      return success;
-    }
-  }
-
-  for (const std::shared_ptr<ChronosInstance>& inst : _chronos_instances)
-  {
-    success = inst->wait_for_instance();
-
-    if (!success)
-    {
-      return success;
-    }
-  }
-
-  return success;
-}
-
-std::vector<std::string> DbSite::get_chronos_ips()
-{
-  std::vector<std::string> ips;
-  for (const std::shared_ptr<ChronosInstance> instance : _chronos_instances)
-  {
-    ips.push_back(instance->ip());
-  }
-
-  return ips;
-}
-
-std::vector<std::string> DbSite::get_rogers_ips()
-{
-  std::vector<std::string> ips;
-  for (const std::shared_ptr<RogersInstance> instance : _rogers_instances)
-  {
-    ips.push_back(instance->ip());
-  }
-
-  return ips;
-}
 
 /// TODO
 ///
